wip: shard reader
